Move enum-collection-test to a separate test
